+++ pytest_anki/__init__.py
@@ -35,9 +35,9 @@ A simple pytest plugin for testing Anki add-ons
 
 __all__ = ["AnkiStateUpdate", "AnkiWebViewType", "AnkiSessionError", "AnkiSession"]
 
-from ._anki import AnkiStateUpdate, AnkiWebViewType  # noqa: F401
-from ._errors import AnkiSessionError  # noqa: F401
-from ._session import AnkiSession  # noqa: F401
+from ._plugin.anki import AnkiStateUpdate, AnkiWebViewType  # noqa: F401
+from ._plugin.errors import AnkiSessionError  # noqa: F401
+from ._plugin.session import AnkiSession  # noqa: F401
 
 __version__ = "2.0.0-dev.1"
 __author__ = "Aristotelis P. (Glutanimate), Michal Krassowski"
