+++ pytest_anki/_plugin/__init__.py
@@ -1,0 +1,6 @@
+"""
+Machinery for creating pytest-anki's pytest fixtures
+
+NOTE: This module is not part of pytest-anki's public API. Its interface might
+change between releases, so you should not rely on it.
+"""
